cache deps of compiler and launcher
