+++ deno1.2.0/deps.js
@@ -1,0 +1,7 @@
+export {move, exists} from 'https://deno.land/std@0.63.0/fs/mod.ts';
+export {readLines} from 'https://deno.land/std@0.63.0/io/bufio.ts';
+export {
+  join,
+  fromFileUrl,
+  dirname
+} from 'https://deno.land/std@0.63.0/path/mod.ts';
