+++ deno1.2.0/lib/launcher.js
@@ -1,10 +1,11 @@
+import {readLines, exists} from '../deps.js';
 const output = await Deno.open('/dev/fd/3', {read: false, write: true});
 
-for await (const line of Deno.iter(Deno.stdin)) {
+for await (const line of readLines(Deno.stdin)) {
   let response = {error: "couldn't execute your function"};
 
   try {
-    const input = JSON.parse(new TextDecoder().decode(line).trim());
+    const input = JSON.parse(line);
     let payload = {};
     for (const [key, value] of Object.entries(input)) {
       if (key === 'value') {
@@ -16,10 +17,23 @@ for await (const line of Deno.iter(Deno.stdin)) {
       }
     }
 
-    const {default: main} = await import('./ow_bundle.js');
-    response = await main(payload);
+    if (await exists('./ow_bundle.js')) {
+      const {default: main} = await import('./ow_bundle.js');
+      response = await main(payload);
+      if (Object.prototype.toString.call(response) !== '[object Object]') {
+        response = {
+          error: 'response returned by the function is not an object'
+        };
+        console.error(response);
+      }
+    } else {
+      response = {
+        error:
+          "couldn't find the bundled file. There might be an error during bundling."
+      };
+    }
   } catch (error) {
-    Deno.stderr.write(new TextEncoder().encode(error.message));
+    console.error(error);
     response = {
       error: error.message
     };
